refactor(bridge): extract message census shape helpers
